+++ platform_sdl.c
@@ -10,6 +10,7 @@
 #include <SDL3/SDL.h>
 #include <stdio.h>
 #include <string.h>
+#include <stdint.h>
 
 /* ── Internal state ────────────────────────────────────────────────── */
 static SDL_Window   *window   = NULL;
@@ -119,7 +120,7 @@ void platform_render(const GameState *gs)
         int tex_x = (int)(gs->hits[x].wall_x * TEX_SIZE);
         if (tex_x >= TEX_SIZE) tex_x = TEX_SIZE - 1;
 
-        int wt   = gs->hits[x].wall_type;
+        uint16_t wt = gs->hits[x].wall_type;
         int side = gs->hits[x].side;
 
         /* Clamp visible range to screen */
