+++ textures_sdl.h
@@ -2,6 +2,7 @@
 #define TEXTURES_SDL_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 /* ── Texture constants ────────────────────────────────────────────── */
 #define TEX_SIZE  64             /* width & height of one texture tile */
@@ -22,6 +23,6 @@ void tm_shutdown(void);
  *   wall_type: 0 .. TEX_COUNT-1
  *   tex_x, tex_y: 0 .. TEX_SIZE-1
  *   Returns colour in RGBA8888 format. */
-unsigned int tm_get_pixel(int wall_type, int tex_x, int tex_y);
+unsigned int tm_get_pixel(uint16_t wall_type, int tex_x, int tex_y);
 
 #endif /* TEXTURES_SDL_H */
